@@ -1,10 +1,23 @@
-import { inject } from '@loopback/core'
-import { DefaultCrudRepository } from '@loopback/repository'
-import { HyponosDataSource } from '../datasources'
-import { Hotel, HotelRelations } from '../models'
+import { Getter, inject } from '@loopback/core'
+import { DefaultCrudRepository, HasManyRepositoryFactory, repository } from '@loopback/repository'
+
+import { PostgresDataSource } from '../datasources'
+import { Hotel, HotelRelations, Room } from '../models'
+import { RoomRepository } from './room.repository'
 
 export class HotelRepository extends DefaultCrudRepository<Hotel, typeof Hotel.prototype.id, HotelRelations> {
-    constructor(@inject('datasources.mongo') dataSource: HyponosDataSource) {
+    public readonly rooms: HasManyRepositoryFactory<Room, typeof Room.prototype.id>
+
+    constructor(
+        @inject('datasources.postgres')
+        dataSource: PostgresDataSource,
+        @repository.getter('RoomRepository')
+        roomRepositoryGetter: Getter<RoomRepository>
+    ) {
         super(Hotel, dataSource)
+
+        this.rooms = this.createHasManyRepositoryFactoryFor('rooms', roomRepositoryGetter)
+
+        this.registerInclusionResolver('rooms', this.rooms.inclusionResolver)
     }
 }
