@@ -3,15 +3,19 @@ import crypto from 'crypto'
 import { BootMixin } from '@loopback/boot'
 import { ApplicationConfig, BindingKey, createBindingFromClass } from '@loopback/core'
 import { RestExplorerBindings, RestExplorerComponent } from '@loopback/rest-explorer'
-import { RepositoryMixin } from '@loopback/repository'
+import { RepositoryMixin, SchemaMigrationOptions } from '@loopback/repository'
 import { RestApplication } from '@loopback/rest'
 import { ServiceMixin } from '@loopback/service-proxy'
 import path from 'path'
+import fs from 'fs'
+import YAML from 'yaml'
 import { AuthenticationComponent } from '@loopback/authentication'
 import { JWTAuthenticationComponent, TokenServiceBindings } from '@loopback/authentication-jwt'
 import { AuthorizationComponent } from '@loopback/authorization'
 
-import { PasswordHasherBindings, UserServiceBindings } from './utils'
+import { UserWithPassword } from './models'
+import { UserRepository } from './repositories'
+import { PasswordHasherBindings, UserServiceBindings } from './utils/keys'
 import { MySequence } from './sequence'
 import { BcryptHasher, JWTService, UserManagementService, SecuritySpecEnhancer } from './services'
 
@@ -74,5 +78,29 @@ export class HyponosApplication extends BootMixin(ServiceMixin(RepositoryMixin(R
         // otherwise create a random string of 64 hex digits
         const secret = process.env.JWT_SECRET ?? crypto.randomBytes(32).toString('hex')
         this.bind(TokenServiceBindings.TOKEN_SECRET).to(secret)
+    }
+
+    async migrateSchema(options?: SchemaMigrationOptions) {
+        await super.migrateSchema(options)
+
+        const userRepository = await this.getRepository(UserRepository)
+        const foundUser = await userRepository.findOne({ where: { email: 'hubertlucas41@gmail.com' } })
+
+        if (!foundUser) {
+            const usersDir = path.join(__dirname, '../fixtures/users')
+            const userFiles = fs.readdirSync(usersDir)
+
+            for (const file of userFiles) {
+                if (file.endsWith('.yml')) {
+                    const userFile = path.join(usersDir, file)
+                    const yamlString = YAML.parse(fs.readFileSync(userFile, 'utf8'))
+                    const userWithPassword = new UserWithPassword(yamlString)
+                    const userManagementService = await this.get<UserManagementService>(
+                        UserServiceBindings.USER_SERVICE
+                    )
+                    await userManagementService.createUser(userWithPassword)
+                }
+            }
+        }
     }
 }
