@@ -1,9 +1,9 @@
 import { authenticate } from '@loopback/authentication'
 import { authorize } from '@loopback/authorization'
-import { Filter, FilterExcludingWhere, repository } from '@loopback/repository'
+import { Filter, repository } from '@loopback/repository'
 import { post, param, get, getModelSchemaRef, patch, del, requestBody, response } from '@loopback/rest'
 
-import { ROLE_ADMIN } from '../constants'
+import { ROLE_ADMIN, ROLE_MANAGER } from '../constants'
 import { Hotel } from '../models'
 import { HotelRepository } from '../repositories'
 import { basicAuthorization } from '../services'
@@ -65,17 +65,14 @@ export class HotelController {
             }
         }
     })
-    async findById(
-        @param.path.string('id') id: string,
-        @param.filter(Hotel, { exclude: 'where' }) filter?: FilterExcludingWhere<Hotel>
-    ): Promise<Hotel> {
-        return this.hotelRepository.findById(id, filter)
+    async findById(@param.path.string('id') id: string): Promise<Hotel> {
+        return this.hotelRepository.findById(id, { include: ['rooms'] })
     }
 
     @patch('/admin/hotels/{id}')
     @authenticate('jwt')
     @authorize({
-        allowedRoles: [ROLE_ADMIN],
+        allowedRoles: [ROLE_ADMIN, ROLE_MANAGER],
         voters: [basicAuthorization]
     })
     @response(204, {
