@@ -1,17 +1,17 @@
 import { SecurityBindings, UserProfile, securityId } from '@loopback/security'
 import _ from 'lodash'
 import { inject } from '@loopback/core'
-import { Count, CountSchema, FilterExcludingWhere, repository, Where } from '@loopback/repository'
+import { Count, CountSchema, Filter, repository, Where } from '@loopback/repository'
 import { post, param, get, getModelSchemaRef, patch, del, requestBody, response, HttpErrors } from '@loopback/rest'
 import { authenticate, TokenService } from '@loopback/authentication'
 import { authorize } from '@loopback/authorization'
 import { TokenServiceBindings } from '@loopback/authentication-jwt'
 
-import { ROLE_ADMIN, ROLE_USER } from '../constants'
+import { ROLE_ADMIN, ROLE_USER, UNIQUE_VIOLATION } from '../constants'
 import { User, UserWithPassword } from '../models'
 import { Credentials, UserRepository } from '../repositories'
 import { basicAuthorization, validateCredentials, UserManagementService } from '../services'
-import { UserServiceBindings } from '../utils'
+import { UserServiceBindings } from '../utils/keys'
 import { CredentialsRequestBody } from './specs/user-controller.specs'
 
 export class UserController {
@@ -49,7 +49,7 @@ export class UserController {
         try {
             return await this.userManagementService.createUser(user)
         } catch (error) {
-            if (error.code === 11000 && error.errmsg.includes('index: uniqueEmail')) {
+            if (error.code === UNIQUE_VIOLATION && error.constraint === 'uniqueEmail') {
                 throw new HttpErrors.Conflict('Cette adresse email existe déjà')
             } else {
                 throw error
@@ -133,11 +133,12 @@ export class UserController {
     })
     async getCurrentUser(
         @inject(SecurityBindings.USER)
-        currentUserProfile: UserProfile
+        currentUserProfile: UserProfile,
+        @param.filter(User) filter?: Filter<User>
     ): Promise<User> {
         const userId = currentUserProfile[securityId]
 
-        return this.userRepository.findById(userId)
+        return this.userRepository.findById(userId, filter)
     }
 
     @get('/admin/users')
@@ -177,7 +178,7 @@ export class UserController {
         return this.userRepository.updateAll(user, where)
     }
 
-    @get('/users/{id}')
+    @get('/admin/users/{id}')
     @response(200, {
         description: 'User model instance',
         content: {
@@ -186,12 +187,8 @@ export class UserController {
             }
         }
     })
-    async findById(
-        @param.path.string('id') id: string,
-        @param.filter(User, { exclude: 'where' })
-        filter?: FilterExcludingWhere<User>
-    ): Promise<User> {
-        return this.userRepository.findById(id, filter)
+    async findById(@param.path.string('id') id: string): Promise<User> {
+        return this.userRepository.findById(id, { include: ['hotels'] })
     }
 
     @patch('/admin/users/{id}')
@@ -210,6 +207,10 @@ export class UserController {
         user: User
     ): Promise<void> {
         const currentUser = await this.userRepository.findById(id)
+
+        if (currentUser.email === 'hello@studi.fr') {
+            throw new HttpErrors.Unauthorized('Vous ne pouvez pas modifier ce compte')
+        }
 
         if (currentUser.role === ROLE_ADMIN) {
             const usersAdmin = await this.userRepository.find({ where: { role: ROLE_ADMIN } })
@@ -232,6 +233,10 @@ export class UserController {
     })
     async deleteById(@param.path.string('id') id: string): Promise<void> {
         const currentUser = await this.userRepository.findById(id)
+
+        if (currentUser.email === 'hello@studi.fr') {
+            throw new HttpErrors.Unauthorized('Vous ne pouvez pas supprimer ce compte')
+        }
 
         if (currentUser.role === ROLE_ADMIN) {
             const usersAdmin = await this.userRepository.find({ where: { role: ROLE_ADMIN } })
