@@ -2,15 +2,14 @@ import { inject, lifeCycleObserver, LifeCycleObserver } from '@loopback/core'
 import { juggler } from '@loopback/repository'
 
 const config = {
-    name: 'mongo',
-    connector: 'mongodb',
+    name: 'postgres',
+    connector: 'postgresql',
+    url: '',
     host: process.env.DB_HOST,
-    port: 27017,
+    port: 5432,
     user: process.env.DB_USER,
     password: process.env.DB_PASSWORD,
-    database: process.env.DB_NAME,
-    useNewUrlParser: true,
-    protocol: 'mongodb+srv'
+    database: process.env.DB_NAME
 }
 
 // Observe application's life cycle to disconnect the datasource when
@@ -18,12 +17,12 @@ const config = {
 // gracefully. The `stop()` method is inherited from `juggler.DataSource`.
 // Learn more at https://loopback.io/doc/en/lb4/Life-cycle.html
 @lifeCycleObserver('datasource')
-export class HyponosDataSource extends juggler.DataSource implements LifeCycleObserver {
-    static dataSourceName = 'mongo'
+export class PostgresDataSource extends juggler.DataSource implements LifeCycleObserver {
+    static dataSourceName = 'postgres'
     static readonly defaultConfig = config
 
     constructor(
-        @inject('datasources.config.mongo', { optional: true })
+        @inject('datasources.config.postgres', { optional: true })
         dsConfig: object = config
     ) {
         super(dsConfig)
