@@ -1,7 +1,6 @@
-import { model, property } from '@loopback/repository'
+import { property } from '@loopback/repository'
 import { User } from './user.model'
 
-@model()
 export class UserWithPassword extends User {
     @property({
         type: 'string',
