@@ -1,7 +1,15 @@
 import { inject, Getter } from '@loopback/core'
-import { DefaultCrudRepository, HasOneRepositoryFactory, repository } from '@loopback/repository'
-import { HyponosDataSource } from '../datasources'
-import { User, UserCredentials } from '../models'
+import {
+    DefaultCrudRepository,
+    HasManyThroughRepositoryFactory,
+    HasOneRepositoryFactory,
+    repository
+} from '@loopback/repository'
+
+import { PostgresDataSource } from '../datasources'
+import { Hotel, User, UserCredentials, UserHotel } from '../models'
+import { HotelRepository } from './hotel.repository'
+import { UserHotelRepository } from './user-hotel.repository'
 import { UserCredentialsRepository } from './user-credentials.repository'
 
 export type Credentials = {
@@ -11,14 +19,32 @@ export type Credentials = {
 
 export class UserRepository extends DefaultCrudRepository<User, typeof User.prototype.id> {
     public readonly userCredentials: HasOneRepositoryFactory<UserCredentials, typeof User.prototype.id>
+    public readonly hotels: HasManyThroughRepositoryFactory<
+        Hotel,
+        typeof Hotel.prototype.id,
+        UserHotel,
+        typeof User.prototype.id
+    >
 
     constructor(
-        @inject('datasources.mongo') dataSource: HyponosDataSource,
+        @inject('datasources.postgres')
+        dataSource: PostgresDataSource,
         @repository.getter('UserCredentialsRepository')
-        protected userCredentialsRepositoryGetter: Getter<UserCredentialsRepository>
+        protected userCredentialsRepositoryGetter: Getter<UserCredentialsRepository>,
+        @repository.getter('HotelRepository')
+        hotelRepositoryGetter: Getter<HotelRepository>,
+        @repository.getter('UserHotelRepository')
+        userHotelRepositoryGetter: Getter<UserHotelRepository>
     ) {
         super(User, dataSource)
         this.userCredentials = this.createHasOneRepositoryFactoryFor('userCredentials', userCredentialsRepositoryGetter)
+        this.hotels = this.createHasManyThroughRepositoryFactoryFor(
+            'hotels',
+            hotelRepositoryGetter,
+            userHotelRepositoryGetter
+        )
+
+        this.registerInclusionResolver('hotels', this.hotels.inclusionResolver)
     }
 
     async findCredentials(userId: typeof User.prototype.id): Promise<UserCredentials | undefined> {
