@@ -7,7 +7,7 @@ import _ from 'lodash'
 
 import { User, UserWithPassword } from '../models'
 import { Credentials, UserRepository } from '../repositories'
-import { PasswordHasherBindings } from '../utils'
+import { PasswordHasherBindings } from '../utils/keys'
 import { PasswordHasher } from './password-hasher'
 
 export class UserManagementService implements UserService<User, Credentials> {
@@ -62,7 +62,6 @@ export class UserManagementService implements UserService<User, Credentials> {
         const password = await this.passwordHasher.hashPassword(userWithPassword.password)
         userWithPassword.password = password
         const user = await this.userRepository.create(_.omit(userWithPassword, 'password'))
-        user.id = user.id.toString()
         await this.userRepository.userCredentials(user.id).create({ password })
         return user
     }
