@@ -1,23 +1,15 @@
-import { HOME, LOGIN } from '@Constants/routes'
 import { ReactNode, createContext, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 
+import { IUser } from '@Interfaces/user'
 import { TOKEN_KEY } from '@Constants/request'
+import { HOME, LOGIN } from '@Constants/routes'
 import UserService from '@Services/user'
-
-export interface IUser {
-    id: string
-    email: string
-    password: string
-    role: string
-    firstname?: string
-    lastname?: string
-}
 
 interface IAuthContextType {
     user: null | IUser
     login: <B extends boolean>(userToken: string, returnUser?: B) => Promise<B extends true ? IUser : void>
-    // signout: (callback: VoidFunction) => void
+    logout: () => void
 }
 
 export const AuthContext = createContext<IAuthContextType>(null!)
@@ -27,7 +19,7 @@ export const AuthProvider = ({ children }: { children: ReactNode }) => {
     const [user, setUser] = useState<null | IUser>(null)
 
     const login = (userToken: string, returnUser = false) => {
-        return UserService.me(userToken)
+        return UserService.me({ token: userToken })
             .then(response => {
                 setUser(response.data)
 
@@ -45,14 +37,13 @@ export const AuthProvider = ({ children }: { children: ReactNode }) => {
             })
     }
 
-    // const logout = (callback: VoidFunction) => {
-    //     return UserService.logout().then(() => {
-    //         setUser(null)
-    //         callback()
-    //     })
-    // }
+    const logout = () => {
+        setUser(null)
+        localStorage.removeItem(TOKEN_KEY)
+        navigate(HOME, { replace: true })
+    }
 
-    const value = { user, login }
+    const value = { user, login, logout }
 
     return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>
 }
