@@ -1,5 +1,5 @@
 import { Button, Grid, TextField, Typography } from '@mui/material'
-import { useEffect, useState } from 'react'
+import { useContext, useEffect, useState } from 'react'
 import { yupResolver } from '@hookform/resolvers/yup'
 import { Helmet } from 'react-helmet-async'
 import { Controller, SubmitHandler, useForm } from 'react-hook-form'
@@ -7,11 +7,13 @@ import { useNavigate, useParams } from 'react-router-dom'
 import * as yup from 'yup'
 import { useSnackbar } from 'notistack'
 
+import { ROLE_ADMIN, ROLE_MANAGER } from '@Constants/roles'
 import { ADMIN_HOTELS } from '@Constants/routes'
 import { IFormInputs } from '@Interfaces/hotel'
 import { DEFAULT_ERROR_MESSAGE, IS_REQUIRED, MIN_CHAR } from '@Constants/form'
 import { TOKEN_KEY } from '@Constants/request'
-import HotelService from '@Services/hotel'
+import { HotelService, UserService } from '@Services/index'
+import { AuthContext } from '@Src/AuthProvider'
 
 const schema = yup
     .object({
@@ -25,6 +27,8 @@ const schema = yup
     .required()
 
 const AdminHotelsEdit = () => {
+    const token = localStorage.getItem(TOKEN_KEY)
+    const auth = useContext(AuthContext)
     const navigate = useNavigate()
     const { hotelId } = useParams()
     const [error, setError] = useState<string | null>(null)
@@ -41,8 +45,6 @@ const AdminHotelsEdit = () => {
         }
     })
     const onSubmit: SubmitHandler<IFormInputs> = data => {
-        const token = localStorage.getItem(TOKEN_KEY)
-
         if (token && hotelId) {
             setError(null)
             HotelService.edit(token, hotelId, data)
@@ -61,17 +63,32 @@ const AdminHotelsEdit = () => {
 
     useEffect(() => {
         if (hotelId) {
-            HotelService.getOne(hotelId)
-                .then(hotel => {
-                    reset(hotel.data)
-                })
-                .catch(error => {
-                    if (error.response.status === 404) {
-                        navigate(ADMIN_HOTELS, { replace: true })
-                    } else {
-                        enqueueSnackbar(DEFAULT_ERROR_MESSAGE, { variant: 'error' })
-                    }
-                })
+            if (auth.user?.role === ROLE_ADMIN) {
+                HotelService.getOne(hotelId)
+                    .then(hotel => {
+                        reset(hotel.data)
+                    })
+                    .catch(error => {
+                        if (error.response.status === 404) {
+                            navigate(ADMIN_HOTELS, { replace: true })
+                        } else {
+                            enqueueSnackbar(DEFAULT_ERROR_MESSAGE, { variant: 'error' })
+                        }
+                    })
+            } else if (auth.user?.role === ROLE_MANAGER && token) {
+                UserService.me({ token, hotelId })
+                    .then(response => {
+                        const hotels = response.data.hotels
+                        if (hotels.length === 0) {
+                            navigate(ADMIN_HOTELS, { replace: true })
+                        } else {
+                            reset(hotels[0])
+                        }
+                    })
+                    .catch(error => {
+                        console.log(error)
+                    })
+            }
         } else {
             navigate(ADMIN_HOTELS, { replace: true })
         }
@@ -187,7 +204,7 @@ const AdminHotelsEdit = () => {
                                 <TextField // TODO replace by WYSIWYG
                                     fullWidth
                                     multiline
-                                    label="Description (optionel)"
+                                    label="Description (optionnel)"
                                     error={invalid}
                                     helperText={error?.message}
                                     rows={4}
