@@ -1,5 +1,7 @@
+import { DEFAULT_ERROR_MESSAGE } from '@Constants/form'
+import { ROLE_ADMIN, ROLE_MANAGER } from '@Constants/roles'
 import { Button, Grid, IconButton, Table, TableBody, TableCell, TableHead, TableRow } from '@mui/material'
-import { useEffect, useState } from 'react'
+import { useContext, useEffect, useState } from 'react'
 import { Helmet } from 'react-helmet-async'
 import { Add, Delete, Edit, Info } from '@mui/icons-material'
 import { Link } from 'react-router-dom'
@@ -8,14 +10,16 @@ import { useSnackbar } from 'notistack'
 import { TOKEN_KEY } from '@Constants/request'
 import { ADMIN_HOTEL, ADMIN_HOTELS_ADD, ADMIN_HOTELS_EDIT } from '@Constants/routes'
 import BasicDialog from '@Components/BasicDialog'
-import HotelService from '@Services/hotel'
+import { HotelService, UserService } from '@Services/index'
 import { IHotel } from '@Interfaces/hotel'
+import { AuthContext } from '@Src/AuthProvider'
 
 const AdminHotels = () => {
     const token = localStorage.getItem(TOKEN_KEY)
+    const auth = useContext(AuthContext)
     const [openDeleteDialog, setOpenDeleteDialog] = useState<boolean>(false)
     const [currentHotelId, setCurrentHotelId] = useState<string | null>(null)
-    const [hotels, setHotels] = useState<null | IHotel[]>(null)
+    const [hotels, setHotels] = useState<null | IHotel[] | []>(null)
     const { enqueueSnackbar } = useSnackbar()
 
     const deleteHotel = () => {
@@ -38,9 +42,17 @@ const AdminHotels = () => {
     }
 
     const fetchHotels = () => {
-        HotelService.getAll()
-            .then(response => setHotels(response.data))
-            .catch(error => console.log(error))
+        if (auth.user?.role === ROLE_ADMIN) {
+            HotelService.getAll()
+                .then(response => setHotels(response.data))
+                .catch(() => {
+                    enqueueSnackbar(DEFAULT_ERROR_MESSAGE, { variant: 'error' })
+                })
+        } else if (auth.user?.role === ROLE_MANAGER && token) {
+            UserService.me({ token, withHotels: true }).then(response => {
+                setHotels(response.data.hotels)
+            })
+        }
     }
 
     useEffect(() => {
@@ -54,13 +66,15 @@ const AdminHotels = () => {
                 <meta name="robots" content="none" />
             </Helmet>
 
-            <Grid container justifyContent="end">
-                <Grid item>
-                    <Button variant="contained" component={Link} to={ADMIN_HOTELS_ADD} startIcon={<Add />}>
-                        Ajouter un hôtel
-                    </Button>
+            {auth.user?.role === ROLE_ADMIN && (
+                <Grid container justifyContent="end">
+                    <Grid item>
+                        <Button variant="contained" component={Link} to={ADMIN_HOTELS_ADD} startIcon={<Add />}>
+                            Ajouter un hôtel
+                        </Button>
+                    </Grid>
                 </Grid>
-            </Grid>
+            )}
 
             {hotels && (
                 <Table>
@@ -79,7 +93,9 @@ const AdminHotels = () => {
                         {Object.values(hotels).map(hotel => (
                             <TableRow key={hotel.id}>
                                 <TableCell>
-                                    <Link to={ADMIN_HOTEL.replace(':hotelId', hotel.id)}>{hotel.id}</Link>
+                                    <Button component={Link} to={ADMIN_HOTEL.replace(':hotelId', hotel.id)}>
+                                        {hotel.id}
+                                    </Button>
                                 </TableCell>
                                 <TableCell>{hotel.name}</TableCell>
                                 <TableCell>{hotel.address}</TableCell>
@@ -95,14 +111,16 @@ const AdminHotels = () => {
                                         <Edit />
                                     </IconButton>
 
-                                    <IconButton
-                                        onClick={() => {
-                                            setOpenDeleteDialog(true)
-                                            setCurrentHotelId(hotel.id)
-                                        }}
-                                    >
-                                        <Delete />
-                                    </IconButton>
+                                    {auth.user?.role === ROLE_ADMIN && (
+                                        <IconButton
+                                            onClick={() => {
+                                                setOpenDeleteDialog(true)
+                                                setCurrentHotelId(hotel.id)
+                                            }}
+                                        >
+                                            <Delete />
+                                        </IconButton>
+                                    )}
                                 </TableCell>
                             </TableRow>
                         ))}
