@@ -1,24 +1,47 @@
-import { Add } from '@mui/icons-material'
-import { Button, Grid, Table, TableBody, TableCell, TableHead, TableRow } from '@mui/material'
-import { useEffect, useState } from 'react'
+import { useSnackbar } from 'notistack'
+import { useContext, useEffect, useState } from 'react'
+import { Link, useNavigate } from 'react-router-dom'
+import { Add, Edit } from '@mui/icons-material'
+import { Button, Grid, IconButton, Table, TableBody, TableCell, TableHead, TableRow } from '@mui/material'
 import { Helmet } from 'react-helmet-async'
 
-import { ADMIN_ROOMS_ADD } from '@Constants/routes'
-import RoomService from '@Services/room'
+import { DEFAULT_ERROR_MESSAGE } from '@Constants/form'
+import { TOKEN_KEY } from '@Constants/request'
+import { ROLE_ADMIN, ROLE_MANAGER } from '@Constants/roles'
+import { ADMIN_HOTEL, ADMIN_ROOMS_ADD, ADMIN_ROOMS_EDIT, HOME } from '@Constants/routes'
+import { RoomService, UserHotelService } from '@Services/index'
 import { IRoom } from '@Interfaces/room'
-import { Link } from 'react-router-dom'
+import { AuthContext } from '@Src/AuthProvider'
 
 const Rooms = () => {
+    const navigate = useNavigate()
+    const auth = useContext(AuthContext)
+    const { enqueueSnackbar } = useSnackbar()
     const [rooms, setRooms] = useState<null | IRoom[]>(null)
 
     const fetchRooms = () => {
         RoomService.getAll()
             .then(response => setRooms(response.data))
-            .catch(error => console.log(error))
+            .catch(() => {
+                enqueueSnackbar(DEFAULT_ERROR_MESSAGE, { variant: 'error' })
+            })
     }
 
     useEffect(() => {
-        fetchRooms()
+        const token = localStorage.getItem(TOKEN_KEY)
+        if (auth?.user?.role === ROLE_ADMIN) {
+            fetchRooms()
+        } else if (auth?.user?.role === ROLE_MANAGER && token) {
+            UserHotelService.getHotels(token)
+                .then(response => {
+                    console.log(response)
+                })
+                .catch(() => {
+                    enqueueSnackbar(DEFAULT_ERROR_MESSAGE, { variant: 'error' })
+                })
+        } else {
+            navigate(HOME, { replace: true })
+        }
     }, [])
 
     return (
@@ -42,11 +65,8 @@ const Rooms = () => {
                         <TableRow>
                             <TableCell>ID</TableCell>
                             <TableCell>Titre</TableCell>
-                            {/*<TableCell>Adresse</TableCell>*/}
-                            {/*<TableCell>Code postal</TableCell>*/}
-                            {/*<TableCell>Ville</TableCell>*/}
-                            {/*<TableCell>Pays</TableCell>*/}
-                            {/*<TableCell>Actions</TableCell>*/}
+                            <TableCell>Hôtel</TableCell>
+                            <TableCell>Actions</TableCell>
                         </TableRow>
                     </TableHead>
                     <TableBody>
@@ -54,15 +74,14 @@ const Rooms = () => {
                             <TableRow key={room.id}>
                                 <TableCell>{room.id}</TableCell>
                                 <TableCell>{room.title}</TableCell>
-                                {/*<TableCell>{hotel.address}</TableCell>*/}
-                                {/*<TableCell>{hotel.postCode}</TableCell>*/}
-                                {/*<TableCell>{hotel.city}</TableCell>*/}
-                                {/*<TableCell>{hotel.country}</TableCell>*/}
-                                {/*<TableCell>*/}
-                                {/*    <IconButton component={Link} to={ADMIN_ROOMS_EDIT.replace(':roomId', room.id)}>*/}
-                                {/*        <Edit />*/}
-                                {/*    </IconButton>*/}
-                                {/*</TableCell>*/}
+                                <TableCell>
+                                    <Link to={ADMIN_HOTEL.replace(':hotelId', room.hotelId)}>{room.hotelId}</Link>
+                                </TableCell>
+                                <TableCell>
+                                    <IconButton component={Link} to={ADMIN_ROOMS_EDIT.replace(':roomId', room.id)}>
+                                        <Edit />
+                                    </IconButton>
+                                </TableCell>
                             </TableRow>
                         ))}
                     </TableBody>
